Extract abstract `Boxpiler` class and `IBoxpiler` interface to prep for ASM
